+++ src/TecanScanner/NoRackIdException.php
@@ -1,0 +1,11 @@
+<?php declare(strict_types=1);
+
+namespace Mll\LiquidHandlingRobotics\TecanScanner;
+
+class NoRackIdException extends TecanScanException
+{
+    public function __construct()
+    {
+        parent::__construct('No valid rack ID scanned');
+    }
+}