+++ src/TecanScanner/TecanScanException.php
@@ -1,0 +1,9 @@
+<?php declare(strict_types=1);
+
+namespace Mll\LiquidHandlingRobotics\TecanScanner;
+
+use Exception;
+
+abstract class TecanScanException extends Exception
+{
+}